feat: Add background with parallax, and better camera handling
Add two backgrounds, one for the world and one for the clouds above the world. The clouds
have even more parallax than the background.

Camera is adjusted to not look past the x-axis boundary of the ground, and if the player
goes out of screen on the y-axis, it will catch up with them. Otherwise, it stays centered
on the main level and follows the player on the x-axis.