fix: Allow player to change power by properly changing sprites
To properly change the sprite, we have to re-use the sprite component for the game obj.
While working on that, I also improved the accessors so that we use `this.sprite` to keep
track of the character and power (since it's stored in sprite name as 'char-power'), and
created `this.char` and `this.power` accessors accordingly.