chore: Fix koopa spinning sprites
It will be more convenient to handle the wake-up process if the first frame of the spin
animation is the one that goes with the wake-up frame where the feet are poking out.

Beetles and Spinies also have a spinning animation but they don't poke their feet out, so
we don't need to adjust those.